requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
librosa
numpy
scipy
cloudinary
requests
python-multipart
celery
redis